Reference the chi-squared distribution model.
There's a chi distribution range defined for each input data length
(always power of 2). After input data are read from input file, the
length of the result gets approximated to the closest power of 2 (i.e.
input length 232, closest is 256).

The ranges are generated by test utilities on given quantity of data.

Note, that the ranges used in this commit are only for test purposes and
may be false.