Early implementation of a memory pool.
The memory pool is used by file list to speed up addition of new files
into the data structure. Calling malloc(3) for each new file is too
expensive. It takes too much time to scan a directory with hundred of
files.

The implementation of a memory pool is based on Mohammad Dashti's memory
pool class. See here:

https://github.com/mdashti/HPDS/blob/master/src/utils/pool.h

This commit does not handle cleanups, and allocation of new blocks is
rather excessive.