Accept filenames passed via stdin.
If the first input file is '-', expect data on the standard input and
read until EOF is reached. This feature increases a performance
significantly in scenarios where TCHunt-ng is used in conjunction with
find(1) utility. Instead of creating a process for every file (as it
would be the case with find(1)'s -exec option), the process is created
only once and all the data are piped into it.

So this command:

find / -type f -exec tchuntng -p {} \;

becomes this:

find / -type f | tchuntng -p -